pyproject.toml
[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "protea-method"
version = "0.0.1"
description = "Pure inference: KNN search, feature compute, apply reranker. Standalone library; no FastAPI, no SQLAlchemy."
authors = ["Francisco Miguel Pérez Canales <frapercan1@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "protea_method", from = "src" }]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]

[tool.poetry.dependencies]
# Pinned to 3.12 to match the protea-contracts constraint upstream.
# Broadening to 3.10/3.11 requires lowering protea-contracts first
# (its pyproject pins 3.12 too); deferred to a coordinated bump.
python = ">=3.12,<4.0"
# Hard rule: no FastAPI, no SQLAlchemy, no protea-core.
# This is the slim shipping path for the inference method.
"protea-contracts" = { git = "https://github.com/frapercan/protea-contracts.git", branch = "master" }
numpy = ">=1.24"
pandas = ">=2.0"
lightgbm = ">=4.0"
faiss-cpu = ">=1.7"

[tool.poetry.group.dev.dependencies]
pytest = ">=7"
pytest-cov = ">=4"
ruff = ">=0.1"
mypy = ">=1.7"
pandas-stubs = ">=2.0"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "N", "RUF"]
# N803/N806 (uppercase Q/R/X/y in args and locals): kept off because
# numpy / ML code uses uppercase to mark matrices vs lowercase for
# vectors; renaming hurts readability without a quality benefit.
# RUF002 (Unicode lookalikes in docstrings): scientific text uses `×`,
# `→`, `≤` legitimately; ASCII rewrites would obscure meaning.
ignore = ["E501", "N803", "N806", "RUF002"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_unused_ignores = true
# numpy's generic ``ndarray`` type arguments add a lot of ``Any, Any``
# bloat without runtime benefit; tighten when ``numpy.typing`` stabilises.
disallow_any_generics = false

[[tool.mypy.overrides]]
module = ["lightgbm", "lightgbm.*", "faiss", "faiss.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q --strict-markers"
